--- client/components/NavLink.vue
@@ -1,11 +1,0 @@
-<template>
-    <li class="border-b-2 border-transparent hover:border-primary h-full flex items-center cursor-pointer transition-all duration-200 px-8">
-        {{ text }}
-    </li>
-</template>
-
-<script lang="ts" setup>
-defineProps({
-    text: String,
-})
-</script>
