--- client/components/Navigation.vue
@@ -1,7 +1,0 @@
-<template>
-    <ul class="text-white flex items-center justify-end h-full">
-        <nav-link text="Home"></nav-link>
-        <nav-link text="Playlists"></nav-link>
-        <nav-link text="Artists"></nav-link>
-    </ul>
-</template>
