+++ client/layouts/default.vue
@@ -1,6 +1,6 @@
 <template>
     <div class="w-screen h-screen bg-secondary">
-        <navbar></navbar>
+        <app-navbar></app-navbar>
         <slot />
     </div>
 </template>
